refactor: Move RaptorCostConverter into Raptor so it can be reused
