+++ application/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/DefaultAccessEgress.java
@@ -3,7 +3,7 @@ package org.opentripplanner.routing.algorithm.raptoradapter.transit;
 import java.util.Objects;
 import org.opentripplanner.framework.model.TimeAndCost;
 import org.opentripplanner.raptor.api.model.RaptorConstants;
-import org.opentripplanner.routing.algorithm.raptoradapter.transit.cost.RaptorCostConverter;
+import org.opentripplanner.raptor.api.model.RaptorCostConverter;
 import org.opentripplanner.street.search.state.State;
 
 /**
