+++ raptor/src/test/java/org/opentripplanner/raptor/_data/transit/TestTransfer.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.raptor._data.transit;
 
-import static org.opentripplanner.routing.algorithm.raptoradapter.transit.cost.RaptorCostConverter.toRaptorCost;
+import static org.opentripplanner.raptor.api.model.RaptorCostConverter.toRaptorCost;
 
 import org.opentripplanner.raptor.api.model.RaptorTransfer;
